feat/modals-fix

+++ backend/src/controllers/contributionController.js
@@ -1,4 +1,5 @@
 import { prisma } from "../../prisma/lib/prisma.js";
+import { v4 as uuidv4 } from "uuid";
 
 const contributionController = {
   allContributions: async (_, res) => {
@@ -6,7 +7,7 @@ const contributionController = {
       const financeContribs = await prisma.contribuicao_Financeira.findMany({
         orderBy: { DataContribuicao: "desc" },
         include: {
-          comprovante: { select: { Imagem: true } },
+        comprovante: { select: { IdComprovante: true, Imagem: true } },
           usuario: {
             include: {
               time_usuarios: {
@@ -26,6 +27,7 @@ const contributionController = {
       const foodContribs = await prisma.contribuicao_Alimenticia.findMany({
         orderBy: { DataContribuicao: "desc" },
         include: {
+        comprovante: { select: { IdComprovante: true, Imagem: true } },
           usuario: {
             include: {
               time_usuarios: {
@@ -52,6 +54,7 @@ const contributionController = {
           Gastos: c.Gastos,
           IdContribuicao: c.IdContribuicaoFinanceira,
           TipoDoacao: "Financeira",
+          Meta: c.Meta,
           DataContribuicao: c.DataContribuicao,
           Quantidade: c.Quantidade,
           Fonte: c.Fonte,
@@ -63,6 +66,7 @@ const contributionController = {
           IdContribuicao: c.IdContribuicaoAlimenticia,
           Gastos: c.Gastos,
           TipoDoacao: "Alimenticia",
+          Meta: c.Meta,
           DataContribuicao: c.DataContribuicao,
           Quantidade: c.Quantidade,
           Fonte: c.Fonte,
@@ -106,6 +110,7 @@ const contributionController = {
         orderBy: { DataContribuicao: "desc" },
         include: {
           usuario: true,
+          comprovante: true,
           contribuicoes_alimento: {
             include: {
               alimento: true,
@@ -134,7 +139,7 @@ const contributionController = {
       allContribs.sort(
         (a, b) => new Date(b.DataContribuicao) - new Date(a.DataContribuicao)
       );
-console.log(allContribs);
+      console.log(allContribs);
       res.json(allContribs);
     } catch (err) {
       console.error("Erro ao buscar contribuições por RA:", err);
@@ -217,10 +222,9 @@ console.log(allContribs);
       Meta,
       Gastos,
       Fonte,
-      Valor,
       PesoUnidade,
-      IdAlimento,
       alimentos,
+      uuid,
       Imagem,
     } = req.body;
 
@@ -231,8 +235,6 @@ console.log(allContribs);
     }
 
     try {
-      let contribuicao;
-
       if (TipoDoacao === "Financeira") {
         if (!Quantidade || !Fonte || Gastos === undefined) {
           return res.status(400).json({
@@ -240,22 +242,28 @@ console.log(allContribs);
               "Quantidade, Fonte e Gastos são obrigatórios para doação financeira.",
           });
         }
+
         const resultado = await prisma.$transaction(async (tx) => {
-          const comprovante = await tx.comprovante.create({
-            data: {
-              Imagem: "",
-            },
-          });
+          let comprovanteId = null;
+          if (Imagem) {
+            const comprovante = await tx.comprovante.create({
+              data: {
+                Imagem: Imagem,
+              },
+            });
+            comprovanteId = comprovante.IdComprovante;
+          }
 
           const contribuicao = await tx.contribuicao_Financeira.create({
             data: {
+              uuid: uuidv4(),
               RaUsuario: Number(RaUsuario),
               TipoDoacao,
               Quantidade: Number(Quantidade),
               Meta: Meta ? Number(Meta) : null,
               Gastos: Number(Gastos),
               Fonte,
-              IdComprovante: comprovante.IdComprovante,
+              IdComprovante: comprovanteId,
             },
             include: {
               usuario: {
@@ -271,6 +279,7 @@ console.log(allContribs);
 
           return contribuicao;
         });
+
         return res.status(201).json({
           message: "Contribuição financeira criada com sucesso!",
           data: { ...resultado, TipoDoacao: "Financeira" },
@@ -289,23 +298,9 @@ console.log(allContribs);
         }
 
         const resultado = await prisma.$transaction(async (tx) => {
-          const alimento = alimentos.map((alimento) => {
-            return tx.alimento.create({
-              data: {
-                IdAlimento: Number(alimento.IdAlimento),
-              },
-              include: {
-                contribuicoes_alimento: {
-                  select: {
-                    IdContribuicaoAlimento: alimento.IdContribuicaoAlimento,
-                  },
-                  IdAlimento: true,
-                },
-              },
-            });
-          });
-          contribuicao = await tx.contribuicao_Alimenticia.create({
+          const contribuicao = await tx.contribuicao_Alimenticia.create({
             data: {
+              uuid: uuidv4(),
               RaUsuario: Number(RaUsuario),
               TipoDoacao,
               Quantidade: Number(Quantidade),
@@ -316,19 +311,49 @@ console.log(allContribs);
               IdAlimento:
                 alimentos.length === 1 ? Number(alimentos[0].IdAlimento) : null,
             },
-            include: {
-              usuario: true,
-            },
           });
+
+          if (alimentos && alimentos.length > 0) {
+            console.log(
+              ` Criando ${alimentos.length} relações com alimentos...`
+            );
+
+            await Promise.all(
+              alimentos.map((alimento) => {
+                console.log("Criando relação:", {
+                  IdContribuicaoAlimenticia:
+                    contribuicao.IdContribuicaoAlimenticia,
+                  IdAlimento: Number(alimento.IdAlimento),
+                });
+
+                return tx.contribuicao_Alimento.create({
+                  data: {
+                    IdContribuicaoAlimenticia:
+                      contribuicao.IdContribuicaoAlimenticia,
+                    IdAlimento: Number(alimento.IdAlimento),
+                  },
+                });
+              })
+            );
+          }
+
+          return contribuicao;
         });
 
         const contribuicaoCompleta =
           await prisma.contribuicao_Alimenticia.findUnique({
             where: {
-              IdContribuicaoAlimenticia: contribuicao.IdContribuicaoAlimenticia,
+              IdContribuicaoAlimenticia: resultado.IdContribuicaoAlimenticia,
             },
             include: {
-              usuario: true,
+              usuario: {
+                select: {
+                  RaUsuario: true,
+                  NomeUsuario: true,
+                  EmailUsuario: true,
+                },
+              },
+              alimento: true,
               contribuicoes_alimento: {
                 include: {
                   alimento: true,
@@ -336,7 +361,6 @@ console.log(allContribs);
               },
             },
           });
-
         return res.status(201).json({
           message: "Contribuição alimentícia criada com sucesso!",
           data: contribuicaoCompleta,
@@ -347,8 +371,7 @@ console.log(allContribs);
         });
       }
     } catch (err) {
-      console.error("Erro ao criar contribuição:", err);
-      res.status(500).json({
+      return res.status(500).json({
         error: "Erro ao criar contribuição.",
         details: err.message,
       });
@@ -359,40 +382,34 @@ console.log(allContribs);
     const { TipoDoacao, IdContribuicao } = req.params;
 
     try {
+      const id = Number(IdContribuicao);
+      if (isNaN(id)) {
+        return res.status(400).json({ error: "ID inválido." });
+      }
+
       let contribuicao;
 
       if (TipoDoacao === "Financeira") {
         contribuicao = await prisma.contribuicao_Financeira.delete({
-          where: { IdContribuicaoFinanceira: Number(IdContribuicao) },
+          where: { IdContribuicaoFinanceira: id },
         });
       } else if (TipoDoacao === "Alimenticia") {
-        await prisma.contribuicao_Alimento.deleteMany({
-          where: { IdContribuicaoAlimenticia: Number(IdContribuicao) },
-        });
-
         contribuicao = await prisma.contribuicao_Alimenticia.delete({
-          where: { IdContribuicaoAlimenticia: Number(IdContribuicao) },
+          where: { IdContribuicaoAlimenticia: id },
         });
       } else {
         return res.status(400).json({
-          error: "Tipo de doação inválido.",
+          error: "Tipo de doação inválido. Use 'Financeira' ou 'Alimenticia'.",
         });
       }
 
-      res.json({
+      return res.status(200).json({
         message: "Contribuição deletada com sucesso!",
         data: contribuicao,
       });
     } catch (err) {
       console.error("Erro ao deletar contribuição:", err);
-
-      if (err.code === "P2025") {
-        return res.status(404).json({
-          error: "Contribuição não encontrada.",
-        });
-      }
-
-      res.status(500).json({
+      return res.status(500).json({
         error: "Erro ao deletar contribuição.",
         details: err.message,
       });
